Fix Backend-Unit-Tests after applying the new auth
