+++ test/backend-unit-tests/spec/config.js
@@ -1,5 +1,6 @@
 'use strict';
 
 module.exports = {
-    db: 'mongodb://localhost/mean_test'
+    db: 'mongodb://localhost/mean_test',
+    secretKey: 'ourTestingSecretKey'
 };
